Fix inverse kinematics. Add unit tests for InverseKinematics
Extend status error id for inverse kinematics to consider the case in which the number of rejected solution is greater than the fixed limit (exit status=2). Optimize algorithm, removing redundant calculation of yErr: this variable is updated together with the solution q and lambda, when the current updated solution is converging to the searched solution (error decrease).
Unit tests, furthermore, are used to validate the function, by means of simply comparing expected output with actual output.